requirements.txt
-i https://pypi.org/simple
cachetools==3.1.1
certifi==2019.6.16
chardet==3.0.4
click==7.0
flask==1.1.1
google-api-core[grpc]==1.13.0
google-auth==1.6.3
google-cloud-iam==0.1.0
google-cloud-kms==1.1.0
googleapis-common-protos[grpc]==1.6.0
grpc-google-iam-v1==0.11.4
grpcio==1.22.0
idna==2.8
itsdangerous==1.1.0
jinja2==2.10.1
markupsafe==1.1.1
protobuf==3.9.0rc1
pyasn1-modules==0.2.5
pyasn1==0.4.5
pytz==2019.1
requests==2.22.0
rsa==4.0
six==1.12.0
urllib3==1.25.3
werkzeug==0.15.4
